+++ pkg/+mic/+interface/+ui/+common/ImageLogical.m
@@ -1,9 +1,7 @@
-classdef ButtonToggle < mic.interface.ui.common.Base
+classdef ImageLogical < mic.interface.ui.common.Base
     
     methods (Abstract)
         
-        % @return {logical 1x1}
-        l = get(this)
         
         % @param {logical 1x1}
         set(this, l)