+++ pkg/+mic/+interface/+ui/+common/Text.m
@@ -1,0 +1,15 @@
+classdef Text < mic.interface.ui.common.Base
+    
+    methods (Abstract)
+        
+        % @return {char 1xm}
+        c = get(this)
+        
+        % @param {char 1xm}
+        set(this, c)
+        
+        
+    end
+    
+end
+