+++ tests/ui/common/test_UiCommonProgressBar.m
@@ -7,5 +7,5 @@ ui = mic.ui.common.ProgressBar();
 
 h = figure();
 ui.build(h, 10, 10);
-ui.setProgress(0.2)
-
+ui.set(0.2)
+ui.get()