+++ tests/ui/common/test_UiCommonToggle.m
@@ -47,3 +47,7 @@ uit = mic.ui.common.Toggle( ...
 fhOnChange = @(src, evt) disp('change');
 addlistener(uit, 'eChange', fhOnChange);
 uit.build(h, 10, 10, 100, 30);
+
+
+
+